refactor: Split Drug and Pharmacy models
