+++ src/models/pharmacy.ts
@@ -1,14 +1,4 @@
-export class Drug {
-  name: string;
-  expiresIn: number;
-  benefit: number;
-
-  constructor(name, expiresIn, benefit) {
-    this.name = name;
-    this.expiresIn = expiresIn;
-    this.benefit = benefit;
-  }
-}
+import { Drug } from "./drug";
 
 export class Pharmacy {
   drugs: Drug[];
